Initial, working example of a MiXCR pipeline
With OAS containing a lot of separate folders (different papers), and each one
of them having a lot of fasta files to align and assemble, it makes sense to
create a Snakemake pipeline with Snakemake.

So far, the pipeline is able to run with a certain directory structure, and
searches for all fasta files within those directories, aligns them, and
assembles them.

Further changes to be implemented:

 - Change the parameters to remove error messages
 - Include proper logging
 - Include file with all the folder names to consider in config.yaml
 - Make sure the pipeline works universally