Included rules for untar and metadata extraction
The files in the nucleotides directory are now automatically untarred,
in order for mixcr analyze to work on the files. Additionally, the
metadata from the json directories is extracted, and saved in a separate
table. Additional small changes to the files, including changing of the
file paths to the Reddy group server.